requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
matplotlib>=3.7.0
seaborn>=0.12.0
yfinance>=0.2.0
requests>=2.31.0
pandas-datareader>=0.10.0
python-dotenv>=1.0.0
